requirements.txt
absl-py==1.4.0
appdirs==1.4.3
apturl==0.5.2
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.11.2
blinker==1.4
Brlapi==0.7.0
cachetools==5.3.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==8.1.3
colorama==0.4.3
command-not-found==0.3
contourpy==1.0.7
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
decorator==5.1.1
deepface==0.0.78
defer==1.0.6
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
dlib==19.24.0
docopt==0.6.2
duplicity==0.8.12.0
easyocr==1.6.2
entrypoints==0.3
executing==1.2.0
fasteners==0.14.1
filelock==3.9.0
fire==0.5.0
Flask==2.2.2
flatbuffers==23.1.21
fonttools==4.38.0
future==0.18.2
gast==0.4.0
gdown==4.6.0
geographiclib==2.0
geopy==2.3.0
gitignore==0.0.8
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-cloud-documentai==2.12.0
google-cloud-vision==3.3.1
google-pasta==0.2.0
googleapis-common-protos==1.58.0
googlemaps==4.10.0
grpcio==1.51.1
grpcio-status==1.51.1
h3==3.7.6
h5py==3.8.0
httplib2==0.14.0
idna==2.8
imageio==2.25.0
importlib-metadata==6.0.0
ipython==8.9.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
karas==0.7.0
keras==2.11.0
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==15.0.6.1
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
monotonic==1.5
more-itertools==4.2.0
mrz==0.6.2
mrz-scanner-sdk==1.0.9
mtcnn==0.1.1
netifaces==0.10.4
networkx==3.0
ninja==1.11.1
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.0
olefile==0.46
opencv-python==4.7.0.68
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
paramiko==2.6.0
parso==0.8.3
pbr==5.11.1
pdf2image==1.16.2
pdf417decoder==1.0.8
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.4.0
pipenv==2023.2.18
platformdirs==2.6.2
prompt-toolkit==3.0.36
proto-plus==1.22.2
protobuf==3.19.6
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pyclipper==1.3.0.post4
pycparser==2.21
pycups==1.9.73
Pygments==2.14.0
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyRFC3339==1.1
PySocks==1.7.1
pytesseract==0.3.10
python-apt==2.0.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.7.1
PyWavelets==1.4.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
retina-face==0.0.13
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
SecretStorage==2.3.1
shapely==2.0.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.3.2.post1
stack-data==0.6.2
stevedore==4.1.1
systemd-python==234
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==1.5
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.1.23.1
timezonefinder==6.1.9
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.9.0
translation==1.0.5
typing_extensions==4.4.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.3
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.14.1
xkit==0.0.0
zipp==3.11.0
zxing==0.14
